[wip] restrucuted server package
